ai/model_server_requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.100.0
uvicorn>=0.23.0
pillow>=9.5.0
numpy>=1.24.0
timm>=0.9.0
python-multipart>=0.0.6
